🎶rename restore map and design spec (#4111)
* rename restore map

* rename spec

* fix build error

* rename to restore map
