+++ src/docfx/lib/github/GitHubUserCache.cs
@@ -93,7 +93,7 @@ namespace Microsoft.Docs.Build
                     return new GitHubUserCache(docset, AppData.DefaultGitHubUserCachePath);
                 }
 
-                var (localPath, content, etag) = await docset.GetRestoredFileContent(path);
+                var (localPath, content, etag) = await RestoreMap.GetRestoredFileContent(docset, path);
                 if (string.IsNullOrEmpty(localPath))
                 {
                     return new GitHubUserCache(docset, path, content, etag, AppData.GetGitHubUserCachePath(path));
